Download archetype metadata and cache it to a workspace
